#3544 – Option to disable "Custom Query" feature
