+++ packages/ketcher-core/src/utilities/index.ts
@@ -20,3 +20,4 @@ export * from './runAsyncAction';
 export * from './b64toBlob';
 export * from './notifyRequestCompleted';
 export * from './KetcherLogger';
+export * from './SettingsManager';
