--- packages/ketcher-react/src/script/ui/utils/optionsManager.ts
@@ -1,41 +1,0 @@
-import { KetcherLogger } from 'ketcher-core';
-import { KETCHER_SAVED_OPTIONS_KEY } from 'src/constants';
-
-interface SavedOptions {
-  ignoreChiralFlag?: boolean;
-}
-
-export class OptionsManager {
-  static getOptions(): SavedOptions {
-    try {
-      return JSON.parse(
-        localStorage.getItem(KETCHER_SAVED_OPTIONS_KEY) || '{}',
-      );
-    } catch (e) {
-      KetcherLogger.error('optionsManager.ts::OptionsManager::getOptions', e);
-      return {} as SavedOptions;
-    }
-  }
-
-  static saveSettings(settings: SavedOptions) {
-    if (!settings) {
-      return;
-    }
-    localStorage.setItem(KETCHER_SAVED_OPTIONS_KEY, JSON.stringify(settings));
-  }
-
-  static get ignoreChiralFlag() {
-    const { ignoreChiralFlag } = this.getOptions();
-
-    return ignoreChiralFlag;
-  }
-
-  static set ignoreChiralFlag(ignoreChiralFlag: boolean | undefined) {
-    const settings = this.getOptions();
-
-    this.saveSettings({
-      ...settings,
-      ignoreChiralFlag,
-    });
-  }
-}
